Add basic health check endpoints
`/health` is an endpoint through which more thorough health checks can
be completed. Currently, it simply confirms that the application is
running, that the ActiveRecord connection is active, and provides a
basic service name.

`/ping` is a simpler endpoint for services like HAProxy and others that
only need a simple response to confirm that the application is
responding. In this case, we use a simple HTTP 200 response.